style: reorganize components
